+++ examples/kbound/omkar/bench/dekker-unsafe1.cpp
@@ -1,7 +1,7 @@
 
-int wants_to_enter[2] = {0,0};
-int turn = 0;
-int var=0;
+long int wants_to_enter[2] = {0,0};
+long int turn = 0;
+long int var=0;
 
 void thread0(){
       //lock
