+++ examples/kbound/omkar/bench/dijkstra-safe.cpp
@@ -1,8 +1,8 @@
 void dmbsy();
 
-int flag[2] = {0,0};
-int turn = 0;
-int var = 0;
+long int flag[2] = {0,0};
+long int turn = 0;
+long int var = 0;
 
 void thread0(){
       begin:
