+++ examples/kbound/omkar/bench/fibbench-unsafe.cpp
@@ -1,13 +1,13 @@
-int i=1,j=1;
+long int i=1,j=1;
 void thread0(){
-      for (int k = 0; k < 5; k++)
+      for (long int k = 0; k < 5; k++)
       {
             i = i+j;
       }
       
 }
 void thread1(){
-      for (int k = 0; k < 5; k++)
+      for (long int k = 0; k < 5; k++)
       {
             j = i+j;
       }
