+++ examples/kbound/omkar/bench/peterson-unsafe1.cpp
@@ -1,7 +1,7 @@
 
-int flag[2];
-int turn;
-int var;
+long int flag[2];
+long int turn;
+long int var;
 
 void thread0(){
       //lock
